+++ src/spellbot/interactions/lfg_interaction.py
@@ -445,11 +445,16 @@ class LookingForGameInteraction(BaseInteraction):
         assert self.ctx.guild
         for player_xid, new_award in new_roles.items():
             if player_xid not in fetched_players:
-                warning = f"Unable to give role {new_award.role} to user <@{player_xid}>"
+                warning = (
+                    f"Unable to {'take' if new_award.remove else 'give'}"
+                    f" role {new_award.role}"
+                    f" {'from' if new_award.remove else 'to'}"
+                    f" user <@{player_xid}>"
+                )
                 await safe_send_channel(self.ctx, warning)
                 continue
             player = fetched_players[player_xid]
-            await safe_add_role(player, self.ctx.guild, new_award.role)
+            await safe_add_role(player, self.ctx.guild, new_award.role, new_award.remove)
             await safe_send_user(player, new_award.message)
 
         # notifiy issues with player permissions